Popravljena greška međuovisnosti, dodane @Lazy notacije da se ne javi greška sa beskonačnom petljom, popravljeno dodavanje Artikla i Usera, više se ne može dodati Clothes ili Footwear objekt direktno nego se doda novi (Article, type) gdje je type "clothes" ili "footwear". Objekti Clothes i Footwear i dalje postoje zajedno sa servisima, samo sam standardizirao. Isto vrijedi sa userima, sada se dodaje (User, role) gdje je role "seller" ili "registereduser". Popravio sam malo i scrapper radi tih promjena i optimizirao ga da brže unosi. Dodan generator ormara koji svakom registriranom useru doda ormar i napuni ga sa random artiklima, trebat ce jos malo doraditi ali radi kak spada. Dodan ispis u terminalu koji obavijesti u kojoj fazi je punjenje baze podataka.

+++ IzvorniKod/backend/src/main/java/com/progi/progi/service/OrmarService.java
@@ -24,18 +24,17 @@ public class OrmarService {
         return (List<Closet>) ormarRepository.findAll();
     }
 
-    public Closet add(String naziv) {
-        Closet closet = new Closet();
-        closet.setClosetname(naziv);
-        closet.setUserid(null);
-        return ormarRepository.save(closet);
-    }
+    public List<Closet> getByUserID(int userID) { return ormarRepository.findByUserID(userID); }
+
+    public Closet add(Closet closet) { return ormarRepository.save(closet); }
 
     public boolean delete(int id) {
         Closet closet = ormarRepository.findById(id).orElse(null);
         List<Location> locations = locationService.getByClosetId(closet.getId());
-        for (Location location : locations) {
-            locationService.delete(location.getId());
+        if (!locations.isEmpty()) {
+            for (Location location : locations) {
+                locationService.delete(location.getId());
+            }
         }
         ormarRepository.delete(closet);
         return true;
